+++ src/services/calcHash.js
