+++ src/services/getOsInfo.js
